requirements.txt
asgiref==3.7.2
bidict==0.22.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==5.0
filelock==3.13.1
Flask==3.0.0
Flask-SocketIO==5.3.6
fsspec==2023.12.2
h11==0.14.0
huggingface-hub==0.17.3
idna==3.6
image==1.5.33
itsdangerous==2.1.2
Jinja2==3.1.2
Levenshtein==0.23.0
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
python-engineio==4.8.0
python-Levenshtein==0.23.0
python-socketio==5.10.0
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
simple-websocket==1.0.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tokenizers==0.14.1
torch==2.1.2
tqdm==4.66.1
transformers==4.34.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
